+++ src/simmate/workflow_engine/execution/cluster/__init__.py
@@ -1,0 +1,4 @@
+# -*- coding: utf-8 -*-
+
+from .local import LocalCluster
+from .slurm import SlurmCluster