+++ docs/change_log.md
@@ -74,6 +74,7 @@ There is one key exception to the rules above -- and that is with `MAJOR`=0 rele
 - add `simmate version` command
 - changelog and update guide added to documentation website
 - add `show-stats`, `delete-finished`, and `delete-all` commands to `workflow-engine`
+- add `Cluster` base class + commands that allow submitting a steady-state cluster via subprocesses or slurm
 
 
 --------------------------------------------------------------------------------